+++ src/app.module.ts
@@ -7,12 +7,14 @@ import { ServeStaticModule } from '@nestjs/serve-static';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { DatabaseModule } from './database/database.module';
+import { StorageUnitModule } from './storage-unit/storage-unit.module';
 
 const frontendDist = join(__dirname, 'frontend');
 
 @Module({
   imports: [
     DatabaseModule,
+    StorageUnitModule,
     ...(existsSync(frontendDist)
       ? [
           ServeStaticModule.forRoot({