+++ src/storage-unit/dto/update-storage-unit.dto.ts
@@ -1,0 +1,5 @@
+import { PartialType } from '@nestjs/mapped-types';
+
+import { CreateStorageUnitDto } from './create-storage-unit.dto';
+
+export class UpdateStorageUnitDto extends PartialType(CreateStorageUnitDto) {}